requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
Pillow
pdf2image
langchain
langchain-openai
langgraph
python-dotenv
typing_extensions
# Add pytest etc. for development if you need